Remove SWAP_INTEGER

+++ variables/discrete/PoissonRand.h
@@ -35,7 +35,7 @@ public:
     double Skewness() const override;
     double ExcessKurtosis() const override;
 
-    bool fitToData(const QVector<double> &sample);
+    bool fitToData(const QVector<int> &sample);
 };
 
 #endif // POISSONRAND_H
